+++ both/controllers/RequestWorkflowController.jsx
@@ -213,22 +213,44 @@ function sendSupplierEmail(request){
       var expiry = moment(request.dueDate).add({days:3}).toDate();
       var token = FMCLogin.generateLoginToken(user,expiry);
       var element = React.createElement(SupplierRequestEmailView,{item:{_id:request._id},token:token});
+      var subject = "New work request from "+" "+team.getName();
       var html = ReactDOMServer.renderToStaticMarkup(element);
 
-      var message = {
-        bcc :["leo@fmclarity.com","rich@fmclarity.com"],
-        from:"FM Clarity <no-reply@fmclarity.com>",
-        subject:("New work request from "+" "+team.getName()),
-        html:html
+      var email, devMsg;
+
+      devMsg = {
+        to:["leo@fmclarity.com","rich@fmclarity.com"]
       }
 
       if(FM.inProduction()) {
-        message.to = to;
+
+        email = {
+          to:to,
+          from:"FM Clarity <no-reply@fmclarity.com>",
+          subject:subject,
+          html:html
+        }
+        Email.send(email);
+
+        devMsg.from = "FM Outgoing Message Alert <no-reply@fmclarity.com>";
+        devMsg.subject = "["+to+"]"+subject;
+        devMsg.html = 
+          "***Message sent to '"+to+"'***<br/><br/>"+
+          html+
+            "<br/>******<br/>"+
+            JSON.stringify(request);
       }
       else {
-        message.subject = "[to:"+to+"]"+message.subject;
+        devMsg.from = "FM Test Message <no-reply@fmclarity.com>";
+        devMsg.subject = "["+to+"]"+subject;
+        devMsg.html = 
+          "***Intercepted message to "+to+"***<br/><br/>"+
+          html+
+            "<br/>******<br/>"+
+            JSON.stringify(request);
       }
-      Email.send(message);
+
+      Email.send(devMsg);      
 
     }
   })}